pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chardet>=5.2.0",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "scrapy>=2.12.0",
    "trafilatura>=2.0.0",
    "wtforms>=3.2.1",
    "python-nmap>=0.7.1",
    "python-whois>=0.9.5",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "flask-wtf>=1.2.2",
    "dnspython>=2.7.0",
    "reportlab>=4.4.0",
    "pymupdf>=1.25.5",
]
